+++ core/common/src/types/message/message_view.rs
@@ -134,10 +134,7 @@ impl<'a> IggyMessageView<'a> {
         let checksum_field_size = size_of::<u64>(); // Skip checksum field for checksum calculation
         let size = self.size() - checksum_field_size;
         let data = &self.buffer[checksum_field_size..checksum_field_size + size];
-
-        // TODO(hubcio): checksum field is 64 bits, but actual checksum is 32 bits (crc32fast)
-        // in future, we should change it to cryptographic safe hash
-        u64::from(checksum::calculate_checksum(data))
+        checksum::calculate_checksum(data)
     }
 }
